Front develop

+++ backend/src/services/programService.ts
@@ -4,10 +4,44 @@ import ProgramStatus from "../enum/programStatus";
 import { calculateProgramStatus } from "../helpers/programHelpers";
 import User from "../models/User";
 import mongoose from "mongoose";
+import UserRoles from "../enum/userRoles";
+
+interface SearchProgramsParams {
+    name?: string;
+    status?: string;
+    country?: string;
+    startDate?: Date;
+    endDate?: Date;
+}
+  
+export const searchPrograms = async (filters: SearchProgramsParams) => {
+    const query: any = {};
+
+    // Añadir filtros dinámicamente al query
+    if (filters.name) {        
+        query.name = { $regex: filters.name, $options: 'i' }; // Búsqueda por nombre (case insensitive)
+    }
+    if (filters.status) {
+        query.status = filters.status;
+    }
+    if (filters.country) {
+        query.country = filters.country;
+    }
+    if (filters.startDate) {
+        query.startDate = { $gte: filters.startDate }; // Programas que inician desde esta fecha
+    }
+    if (filters.endDate) {
+        query.endDate = { $lte: filters.endDate }; // Programas que terminan antes de esta fecha
+    }
+
+    // Ejecutar la consulta y devolver los resultados
+    return await Program.find(query).populate('country participants');
+};
 
 export const createProgram = async (
     name: string,
     countryId: string,
+    participants: string[],
     startDate: Date,
     endDate: Date
 ) => {
@@ -20,9 +54,16 @@ export const createProgram = async (
     if (status === ProgramStatus.COMPLETED) {
         throw new Error('La fecha de inicio no puede ser mayor a la fecha de fin');
     }
+
+    const validParticipants = await User.find({ _id: { $in: participants } });
+    if (validParticipants.length !== participants.length) {
+        throw new Error('Algunos participantes no son válidos');
+    }
+
     const newProgram = new Program({ 
         name,
         country: countryId,
+        participants: validParticipants.map(user => user._id),
         startDate, 
         endDate,
         status
@@ -31,11 +72,32 @@ export const createProgram = async (
     return await newProgram.save();
 }
 
-export const getProgramById = async (programId: string) => {
+export const getProgramById = async (
+    programId: string,
+    userId: string,
+    role: string
+) => { 
+    
     const program = await Program.findById(programId).populate('country participants');
     if (!program) {
         throw new Error('Programa no encontrado');
     }
+    if (role !== UserRoles.Admin) {
+        // Verificar si el usuario está inscrito en el programa
+        const isParticipant = program.participants.some(
+          (participant: any) => participant._id.toString() === userId
+        );
+    
+        // Si el usuario no está inscrito, no devolver participantes
+        if (!isParticipant) {
+          program.participants = []; // No devolver participantes
+        } else {
+          // Si el usuario está inscrito, devolver solo su información
+          program.participants = program.participants.filter(
+            (participant: any) => participant._id.toString() === userId
+          );
+        }
+      }
     return program;
 }
 export const getPrograms = async () => {
@@ -46,12 +108,32 @@ export const updateProgram = async (
     programId: string,
     name?: string,
     startDate?: Date,
-    endDate?: Date
+    endDate?: Date,
+    countryId?: string,
+    participants?: string[],
 ) => {
     const program = await Program.findById(programId);
     if (!program) {
         throw new Error('Programa no encontrado');
     }
+    
+    const country = await Country.findById(countryId);
+    
+    if (!country) {
+        throw new Error('País no encontrado');
+    }
+    if (program.country.toString()!==country.id && program.participants.length > 0) {
+        throw new Error('No se puede modificar el programa porque existen participantes asociados');
+    }
+    if (participants) {
+        const validParticipants = await User.find({ _id: { $in: participants } });
+        if (validParticipants.length !== participants.length) {
+            throw new Error('Algunos participantes no son válidos');
+        }     
+        program.participants = validParticipants.map(user => user._id as mongoose.Types.ObjectId);        
+    }
+
+    if (countryId) program.country = country._id as mongoose.Types.ObjectId;
     if (name) program.name = name;
     if (startDate) program.startDate = startDate;
     if (endDate) program.endDate = endDate;
@@ -63,10 +145,15 @@ export const updateProgram = async (
 }
 
 export const deleteProgram = async (programId: string) => {
-    const program = await Program.findByIdAndDelete(programId);
+    const program = await Program.findById(programId);
     if (!program) {
         throw new Error('Programa no encontrado');
     }
+    if (program.participants.length > 0) {
+        throw new Error('No se puede eliminar el programa porque existen participantes asociados');
+    }
+    return await Program.findByIdAndDelete(programId);
+
 }
 
 export const addParticipant = async (programId: string, userId: string) => {
@@ -112,4 +199,18 @@ export const removeParticipant = async (programId: string, userId: string) => {
         (participantId) => !participantId.equals(userId)
     );
     return await program.save();
+}
+
+export const getByParticipant = async (userId: string) => {
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+        throw new Error(`ID de usuario inválido: ${userId}`);
+    }
+    
+    const user = await User.findById(userId);
+    if (!user) {
+        throw new Error('Usuario no encontrado');
+    }
+    const userObjectId = new mongoose.Types.ObjectId(userId);
+    return await Program.find({ participants: userObjectId }).populate('country participants');
 }
